+++ src/screens/index.js
@@ -6,10 +6,13 @@ export { default as ScreenWaypointBadCondition } from './wpbadcondition';
 export { default as ScreenWaypointCollection } from './wpcollection';
 export { default as ScreenWaypointDashboard } from './wpdashboard';
 export { default as ScreenWaypointGalery } from './wpgalery';
+export { default as ScreenWaypointResume } from './wpresume';
 
 export const NAVS = {
-  start: { next: 'WaypointStack' },
-  driver: { previous: 'ScreenSplash', next: 'ScreenCar' },
-  car: { previous: 'ScreenDriver', next: 'ScreenWaypointDashboard' },
-  wpdashboard: { previous: 'ScreenCar' },
+  start: { current: 'ScreenSplash', next: 'WaypointStack' },
+  driver: { current: 'ScreenDriver', previous: 'ScreenSplash', next: 'ScreenCar' },
+  car: { current: 'ScreenCar', previous: 'ScreenDriver', next: 'ScreenWaypointDashboard' },
+  wpdashboard: { current: 'ScreenWaypointDashboard', previous: 'ScreenCar' },
+  wpbadcondition: { current: 'ScreenWaypointBadCondition' },
+  wpresume: { current: 'ScreenWaypointResume' },
 };